% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check.packages}
\alias{check.packages}
\title{Check package installed}
\usage{
check.packages(pkg)
}
\arguments{
\item{pkg}{package name}
}
\description{
This function check package successfully installed. If not, package install.
}
\keyword{genomic}
