% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_barplot.R
\name{Draw_barplot}
\alias{Draw_barplot}
\title{Draw bar plot using table}
\usage{
Draw_barplot(table)
}
\arguments{
\item{table}{Table for visualizing.}
}
\description{
This function draw bar chart with table
}
\keyword{Genomic}
