% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_pieplot.R
\name{Draw_pieplot}
\alias{Draw_pieplot}
\title{Draw pie plot using table}
\usage{
Draw_pieplot(table)
}
\arguments{
\item{table}{Table for visualizing.}
}
\description{
This function help to draw pie chart by assigning slices and labels.
}
\keyword{Genomic}
