% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_query.R
\name{sql_query}
\alias{sql_query}
\title{SQL query setting}
\usage{
sql_query(sql, input)
}
\arguments{
\item{sql}{sql query}

\item{input}{input data when connecting DB}
}
\description{
This function get sql and extract data.
}
\keyword{genomic}
